:art: Dismatle download button from analysis table component
